Make README.md more inviting to contributors
The README.md previously stated non CODEOWNERS could not contribute. This
is in violation of the spirit of the Istio project. Correct that problem by
making the README.md more inviting.

Our goal is to build a team. The code is a byproduct of that effort.

The MVP was a little difficult to parse and used an ordered list. Simplify the
wording of the MVP such that it is a little more vague, but a little less wordy.
Finally remove the ordering of the MVP feature set since all requirements are
of equivalent value.