DNM: Next-Gen building system
This PR lints/builds/test all code using pinned tooling out of the
istio/tools repository.

This is part of the buildtools standardization effort.  Some of the
original makefile potentially contains cruft.  I could use some help
deteremining what is cruft and what is necessary for correct operation.

The goal of this work is to standardize our toolchain across all
deliverables of Istio.