WIP: Reviews Encouraged - Eject gogo from operator
The operator appears to have a nonstandard CRD interface. This PR
does two things. It standardizes (partially) on the upstream K8s
CRD interface for `IstioControlPlane` and additionally removes
most of gogo. As gogo is a performance tuning system, there is no
strong motivation to use gogo. The usage of gogo is trigging confusing
type conversions in the code base including marshal to translate
values. Further because of bugs in gogo, some funky translation of
go structs and sedding of output variables is happening - which is wrong :)

`Co-Authored-By: Morven Cao <llcao@cn.ibm.com>`

Thanks to @morvencaoMorven for triggering the thinking process
after root causing with his work in PR: https://github.com/istio/operator/pull/381